Fix auto_restart and resume bugs.
Serban is a shit and it made a big mess on resuming from previous state
like it never needs this feature.

Now, pass in --auto_restart and --prefix <path/to/save_dir, without _model.npz>
and we will resume from the previous dump. That is it.
The RUN_ID is also a bug. It messes up all the save_dir. now it is gone.